Split up ifdef'd module code
